v0.5 single-dixcrete-output(multi-players)/requirements.txt
gym==0.20.0
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.2.54
Pillow==9.5.0
PyAutoGUI==0.9.53
pynput==1.7.3
pytesseract==0.3.10
pywin32==303
torch==1.11.0+cu113
tqdm==4.64.0
#asposestorage==1.0.2
easyocr==1.6.2
#mmdet==3.0.0
pyminitouch==0.3.3
#utils==1.0.1